requirements.txt
fastapi==0.96.0
setuptools==67.8.0
SQLAlchemy~=2.0.15
numpy~=1.24.3
nltk~=3.8.1
pandas~=2.0.1
scikit-learn~=1.2.2
networkx~=3.1
pymysql~=1.0.3
uvicorn[standard]==0.22.0
beautifulsoup4~=4.12.2
requests~=2.31.0
python-dotenv~=1.0.0
pydantic~=1.10.8
cryptography==41.0.5
sentry-sdk==1.37.1
pyjwt==2.8.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
